# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI 智能体 API
  description: 基于大模型的智能问答系统，支持翻译、AI文章分析、摘要、推荐、职业分析等功能
  version: 1.0.0
  contact:
    name: AI Agent Team

servers:
  - url: http://localhost:8000
    description: 本地开发服务器

paths:
  /:
    get:
      summary: 根路径
      description: 返回服务基本信息
      operationId: root_get
      tags:
        - 通用
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: AI 智能体 API 服务
                  version:
                    type: string
                    example: 1.0.0
                  docs:
                    type: string
                    example: /docs

  /health:
    get:
      summary: 健康检查
      description: 检查服务是否正常运行
      operationId: health_check_get
      tags:
        - 通用
      responses:
        '200':
          description: 服务健康
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: healthy

  /api/translate:
    post:
      summary: 文本翻译
      description: |
        文本翻译接口。
        语言方向：中文->直接返回原文，非中文->翻译为中文
      operationId: translate_text
      tags:
        - 翻译
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TranslateRequest'
      responses:
        '200':
          description: 翻译成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslateResponse'
        '500':
          description: 翻译失败
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string

  /api/ai-article:
    post:
      summary: AI文章分析
      description: |
        AI文章分类与审核接口。
        - 判断是否为AI相关内容
        - 分类：动态/学习/工具/职业
        - 提取核心要点、摘要和关键词
        - 返回分类置信度和理由
      operationId: analyze_ai_article
      tags:
        - 文章分析
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AIArticleRequest'
      responses:
        '200':
          description: 分析成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIArticleResponse'
        '500':
          description: 分析失败
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string

  /api/wechat-article:
    post:
      summary: 获取微信公众号文章
      description: |
        通过提供微信公众号文章链接，直接获取文章的标题、作者、正文内容、摘要、发布时间等信息。
        支持微信反爬机制处理。
      operationId: get_wechat_article
      tags:
        - 公众号文章
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WechatArticleRequest'
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WechatArticleResponse'
        '500':
          description: 获取失败
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string

  /api/summary:
    post:
      summary: 文本摘要
      description: |
        文本摘要接口，生成简洁准确的摘要内容。
        支持关键词提取和多语言摘要。
      operationId: generate_summary
      tags:
        - 摘要
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SummaryRequest'
      responses:
        '200':
          description: 摘要生成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SummaryResponse'
        '500':
          description: 摘要生成失败
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string

  /api/recommend:
    post:
      summary: 学习路径推荐
      description: |
        基于用户职业和学习目标，推荐学习路径。
        使用RAG模式进行智能推荐。
      operationId: recommend_learning_path
      tags:
        - 推荐
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecommendationRequest'
      responses:
        '200':
          description: 推荐成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationResponse'
        '500':
          description: 推荐失败
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string

  /api/profession/analyze:
    post:
      summary: 职业分析
      description: |
        分析特定职业受AI自动化的影响程度。
        返回风险等级、自动化率、分析内容和建议。
        使用Agent模式进行深度分析。
      operationId: analyze_profession
      tags:
        - 职业分析
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfessionAnalysisRequest'
      responses:
        '200':
          description: 分析成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfessionAnalysisResponse'
        '500':
          description: 分析失败
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string

components:
  schemas:
    TranslateRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          description: 待翻译的文本
          minLength: 1
          example: Hello, world!
        model:
          type: string
          description: 模型标识符：glm, deepseek, qwen, custom
          example: deepseek

    TranslateResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        translated_text:
          type: string
          description: 翻译结果（中文原文或翻译后的中文）

    AIArticleRequest:
      type: object
      required:
        - title
        - content
      properties:
        title:
          type: string
          description: 文章标题
          minLength: 1
          example: OpenAI发布GPT-5新一代大语言模型
        content:
          type: string
          description: 文章内容
          minLength: 10
          example: OpenAI在今日正式发布了GPT-5...
        model:
          type: string
          description: 模型标识符：glm, deepseek, qwen, custom
          example: glm

    AIArticleResponse:
      type: object
      properties:
        category:
          type: string
          description: 分类结果：动态/学习/工具/职业
          example: 动态
        confidence:
          type: number
          format: float
          description: 置信度 (0-1)
          example: 0.85
        reason:
          type: string
          description: 分类理由
          example: 文章报道了AI行业最新动态
        is_ai_related:
          type: boolean
          description: 是否AI相关内容
          example: true
        key_points:
          type: array
          items:
            type: string
          description: 核心要点列表
          example: ["GPT-5发布", "新架构设计", "下月测试"]
        summary:
          type: string
          description: 文章摘要
          example: OpenAI发布GPT-5新一代大模型
        keywords:
          type: array
          items:
            type: string
          description: 关键词列表
          example: ["OpenAI", "GPT-5", "大语言模型"]
        model:
          type: string
          description: 使用的模型
          example: glm

    WechatArticleRequest:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          description: 微信公众号文章链接
          minLength: 1
          example: https://mp.weixin.qq.com/s?__biz=MjM5NTg0NDE1Mw==&mid=2652627839&idx=2&sn=b542d97a02ae6adf771b23a4ba199258

    WechatArticleResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 是否成功获取
          example: true
        title:
          type: string
          description: 文章标题
          example: 深入理解AI Agent技术
        author:
          type: string
          description: 文章作者
          example: 科技前沿
        content:
          type: string
          description: 文章正文内容
          example: 本文深入探讨了AI Agent的核心概念...
        summary:
          type: string
          description: 内容摘要
          example: 本文深入探讨了AI Agent的核心概念...
        publish_time:
          type: string
          description: 发布时间
          example: 2026-04-20
        source_url:
          type: string
          description: 原始链接
          example: https://mp.weixin.qq.com/s/xxx
        cover_image:
          type: string
          description: 封面图片URL
          example: https://mmbiz.qpic.cn/xxx
        error:
          type: string
          description: 错误信息（如果失败）
          example: ""

    SummaryRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          description: 待摘要文本
          example: 长文本内容...
        max_length:
          type: integer
          description: 摘要最大长度
          default: 200
          minimum: 50
          maximum: 1000
        include_keywords:
          type: boolean
          description: 是否提取关键词
          default: true
        language:
          type: string
          description: 摘要语言
          default: 中文

    SummaryResponse:
      type: object
      properties:
        summary:
          type: string
          description: 生成的摘要
          example: 这是文本的摘要内容...
        keywords:
          type: array
          items:
            type: string
          description: 关键词列表
          example: ["AI", "技术", "发展"]
        length:
          type: integer
          description: 摘要长度
          example: 150
        model:
          type: string
          description: 使用的模型
          example: deepseek-chat
        request_id:
          type: string
          description: 请求ID
          example: req_123456

    RecommendationRequest:
      type: object
      required:
        - user_profession
      properties:
        user_profession:
          type: string
          description: 用户职业
          example: 软件工程师
        learning_goals:
          type: array
          items:
            type: string
          description: 学习目标列表
          example: ["提升AI能力", "掌握机器学习"]
        target_days:
          type: integer
          description: 目标天数
          example: 60

    RecommendationResponse:
      type: object
      properties:
        learning_paths:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                example: 入门级AI应用
              duration_days:
                type: integer
                example: 30
              stages:
                type: array
                items:
                  type: object
                  properties:
                    name:
                      type: string
                    content:
                      type: string
                    duration_days:
                      type: integer
              reason:
                type: string
                example: 适合初学者快速入门
        model:
          type: string
          example: deepseek-chat
        request_id:
          type: string
          example: req_123456

    ProfessionAnalysisRequest:
      type: object
      required:
        - profession
      properties:
        profession:
          type: string
          description: 职业名称
          example: 会计
        user_profile:
          type: object
          description: 用户画像
          example:
            experience: 5年
            skills:
              - 财务分析
        include_recommendations:
          type: boolean
          description: 是否包含建议
          default: true

    ProfessionAnalysisResponse:
      type: object
      properties:
        risk_level:
          type: string
          description: 风险等级（高/中/低）
          example: 中
        automation_rate:
          type: number
          format: float
          description: 自动化率 (0-100)
          example: 45.5
        analysis:
          type: string
          description: 分析内容
          example: 该职业部分工作可被自动化替代
        recommendations:
          type: array
          items:
            type: string
          description: 建议列表
          example:
            - 学习数据分析技能
            - 掌握AI辅助工具
        model:
          type: string
          example: deepseek-chat
        request_id:
          type: string
          example: req_123456

tags:
  - name: 通用
    description: 通用接口
  - name: 翻译
    description: 翻译相关接口
  - name: 文章分析
    description: AI文章分析接口（分类、审核、信息提取）
  - name: 公众号文章
    description: 微信公众号文章获取接口
  - name: 摘要
    description: 文本摘要接口
  - name: 推荐
    description: 学习路径推荐接口
  - name: 职业分析
    description: 职业分析相关接口